_cursesmodule.c: Fix is_term_resized() invocation
The PDCurses `is_termresized()` function, which is mapped to
`is_term_resized()`, does not take any arguments.

This was overlooked because VC 14.40 and below did not treat the extra
arguments as an error (only a warning was output).

Note that this does not introduce any behavioural changes since the
extra arguments provided to `is_termresized()` was never used in the
first place.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>